pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "psi-cli"
version = "0.1.0"
description = "CLI tool for psi-agent provenance tracking"
requires-python = ">=3.10"
license = "MIT"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
psi-cli = "psi_cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]
